+++ Tests/RichTextKitTests/RichTextCoordinator+SubscriptionsTests.swift
@@ -96,25 +96,25 @@ final class RichTextCoordinator_SubscriptionsTests: XCTestCase {
 
     func testIsBoldUpdatesTextView() {
         XCTAssertFalse(textView.richTextStyles.hasStyle(.bold))
-        textContext.userActionPublisher.send(.setStyle(.bold, true))
+        textContext.actionPublisher.send(.setStyle(.bold, true))
         XCTAssertTrue(textView.richTextStyles.hasStyle(.bold))
     }
 
     func testIsItalicUpdatesTextView() {
         XCTAssertFalse(textView.richTextStyles.hasStyle(.italic))
-        textContext.userActionPublisher.send(.setStyle(.italic, true))
+        textContext.actionPublisher.send(.setStyle(.italic, true))
         XCTAssertTrue(textView.richTextStyles.hasStyle(.italic))
     }
 
     func testIsUnderlinedUpdatesTextView() {
         XCTAssertFalse(textView.richTextStyles.hasStyle(.underlined))
-        textContext.userActionPublisher.send(.setStyle(.underlined, true))
+        textContext.actionPublisher.send(.setStyle(.underlined, true))
         XCTAssertTrue(textView.richTextStyles.hasStyle(.underlined))
     }
 
     func testIsStrikeThroughUpdatesTextView() {
         XCTAssertFalse(textView.richTextStyles.hasStyle(.strikethrough))
-        textContext.userActionPublisher.send(.setStyle(.strikethrough, true))
+        textContext.actionPublisher.send(.setStyle(.strikethrough, true))
         XCTAssertTrue(textView.richTextStyles.hasStyle(.strikethrough))
     }