+++ Tests/RichTextKitTests/RichTextCoordinatorTests.swift
@@ -64,9 +64,9 @@ final class RichTextCoordinatorTests: XCTestCase {
         let styles = view.richTextStyles
         XCTAssertEqual(context.fontName, view.richTextFont?.fontName)
         XCTAssertEqual(context.fontSize, view.richTextFont?.pointSize)
-        XCTAssertEqual(context.isBold, styles.hasStyle(.bold))
-        XCTAssertEqual(context.isItalic, styles.hasStyle(.italic))
-        XCTAssertEqual(context.isUnderlined, styles.hasStyle(.underlined))
+        XCTAssertEqual(context.styles[.bold], styles.hasStyle(.bold))
+        XCTAssertEqual(context.styles[.italic], styles.hasStyle(.italic))
+        XCTAssertEqual(context.styles[.underlined], styles.hasStyle(.underlined))
         XCTAssertEqual(context.selectedRange, view.selectedRange)
         #if iOS || os(tvOS)
         XCTAssertEqual(context.textAlignment, view.richTextAlignment)